+++ core/execution/spot/auction.go
@@ -91,7 +91,7 @@ func (m *Market) checkAuction(ctx context.Context, now time.Time, idgen common.I
 		// opening auction period has expired, and we have trades, we should be ready to leave
 		// other requirements still need to be checked downstream though
 		m.as.SetReadyToLeave()
-		m.pMonitor.CheckPrice(ctx, m.as, trades, true, false)
+		m.pMonitor.CheckPrice(ctx, m.as, trades, true, true)
 		if m.as.ExtensionTrigger() == types.AuctionTriggerPrice {
 			// this should never, ever happen
 			m.log.Panic("Leaving opening auction somehow triggered price monitoring to extend the auction")